+++ src/Resources/contao/dca/tl_module.php
@@ -12,9 +12,10 @@
 /**
  * Add palette to tl_module
  */
-$GLOBALS['TL_DCA']['tl_module']['palettes']['elo_toplist']  = '{title_legend},name,headline,type;{options_legend},elo_topcount;{protected_legend:hide},protected;{expert_legend:hide},guests,cssID';
-$GLOBALS['TL_DCA']['tl_module']['palettes']['elo_bestlist'] = '{title_legend},name,headline,type;{options_legend},elo_fromdate,elo_todate,elo_min,elo_gender,elo_topcount;{protected_legend:hide},protected;{expert_legend:hide},guests,cssID';
-$GLOBALS['TL_DCA']['tl_module']['palettes']['elo_topx']     = '{title_legend},name,headline,type;{options_legend},elo_topx,elo_gender,elo_fromdate,elo_todate,elo_fidelink;{protected_legend:hide},protected;{expert_legend:hide},guests,cssID';
+$GLOBALS['TL_DCA']['tl_module']['palettes']['elo_toplist']   = '{title_legend},name,headline,type;{options_legend},elo_topcount;{template_legend:hide},customTpl;{protected_legend:hide},protected;{expert_legend:hide},guests,cssID';
+$GLOBALS['TL_DCA']['tl_module']['palettes']['elo_bestlist']  = '{title_legend},name,headline,type;{options_legend},elo_fromdate,elo_todate,elo_min,elo_gender,elo_topcount;{template_legend:hide},customTpl;{protected_legend:hide},protected;{expert_legend:hide},guests,cssID';
+$GLOBALS['TL_DCA']['tl_module']['palettes']['elo_topx']      = '{title_legend},name,headline,type;{options_legend},elo_topx,elo_gender,elo_fromdate,elo_todate,elo_fidelink;{template_legend:hide},customTpl;{protected_legend:hide},protected;{expert_legend:hide},guests,cssID';
+$GLOBALS['TL_DCA']['tl_module']['palettes']['elo_statistik'] = '{title_legend},name,headline,type;{options_legend},elo_liste,elo_statistik;{template_legend:hide},customTpl;{protected_legend:hide},protected;{expert_legend:hide},guests,cssID';
 
 $GLOBALS['TL_DCA']['tl_module']['fields']['elo_topcount'] = array
 (
@@ -89,3 +90,33 @@ $GLOBALS['TL_DCA']['tl_module']['fields']['elo_fidelink'] = array
 	'eval'                               => array('tl_class'=>'w50 clr'),
 	'sql'                                => "char(1) NOT NULL default ''"
 );
+
+$GLOBALS['TL_DCA']['tl_module']['fields']['elo_liste'] = array
+(
+	'label'                              => &$GLOBALS['TL_LANG']['tl_module']['elo_liste'],
+	'exclude'                            => true,
+	'inputType'                          => 'select',
+	'foreignKey'                         => 'tl_elo_listen.title',
+	'eval'                               => array
+	(
+		'includeBlankOption'             => true,
+		'chosen'                         => true,
+		'tl_class'                       => 'w50'
+	),
+	'sql'                                => "int(10) unsigned NOT NULL default 0"
+);
+
+$GLOBALS['TL_DCA']['tl_module']['fields']['elo_statistik'] = array
+(
+	'label'                   => &$GLOBALS['TL_LANG']['tl_module']['elo_statistik'],
+	'exclude'                 => true,
+	'filter'                  => true,
+	'inputType'               => 'checkboxWizard',
+	'options'                 => &$GLOBALS['TL_LANG']['tl_module']['elo_statistik_options'],
+	'eval'                    => array
+	(
+		'multiple'            => true,
+		'tl_class'            => 'w50'
+	),
+	'sql'                     => 'blob NULL'
+);