+++ src/Resources/contao/languages/de/default.php
@@ -17,6 +17,27 @@
  */
 $GLOBALS['TL_LANG']['MSC'][''] = '';
 
-$GLOBALS['TL_LANG']['CTE']['schach'] = "Schach-Elemente";
-$GLOBALS['TL_LANG']['CTE']['eloliste']['0'] = "FIDE-Eloliste";
-$GLOBALS['TL_LANG']['CTE']['eloliste']['1'] = "FIDE-Eloliste einbinden";
+$GLOBALS['TL_LANG']['CTE']['schach'] = 'Schach-Elemente';
+$GLOBALS['TL_LANG']['CTE']['eloliste'] = array('FIDE-Eloliste', 'FIDE-Eloliste einbinden');
+
+
+$GLOBALS['TL_LANG']['tl_module']['elo_statistik_options'] = array
+(
+	'1'  => 'Anzahl aller Spieler',
+	'2'  => 'Anzahl der Spieler mit Elo-Zahl',
+	'3'  => 'Anzahl der Spieler ohne Elo-Zahl',
+	'4'  => 'Anzahl der Spieler mit GM-Titel',
+	'5'  => 'Anzahl der Spieler mit IM-Titel',
+	'6'  => 'Anzahl der Spieler mit FM-Titel',
+	'7'  => 'Anzahl der Spieler mit CM-Titel',
+	'8'  => 'Anzahl der Spieler mit WGM-Titel',
+	'9'  => 'Anzahl der Spieler mit WIM-Titel',
+	'10' => 'Anzahl der Spieler mit WFM-Titel',
+	'11' => 'Anzahl der Spieler mit WCM-Titel',
+	'12' => 'Anzahl der Spieler mit Elo ≥ 2500',
+	'13' => 'Anzahl der Spieler mit Elo ≥ 2400',
+	'14' => 'Anzahl der Spieler mit Elo ≥ 2300',
+	'15' => 'Anzahl der Spieler mit Elo ≥ 2200',
+	'16' => 'Anzahl der Spieler mit Elo ≥ 2100',
+	'17' => 'Anzahl der Spieler mit Elo ≥ 2000',
+);